// src/enums/pixel-channel.ts
/*
  Copyright Dirk Lemstra https://github.com/dlemstra/magick-wasm.
  Licensed under the Apache License, Version 2.0.
*/

/**
 * Specifies the pixel channels.
 */
export enum PixelChannel {
    /**
     * Red.
     */
    Red = 0,

    /**
     * Cyan.
     */
    Cyan = Red,

    /**
     * Gray.
     */
    Gray = Red,

    /**
     * Green.
     */
    Green = 1,

    /**
     * Magenta.
     */
    Magenta = Green,

    /**
     * Blue.
     */
    Blue = 2,

    /**
     * Yellow.
     */
    Yellow = Blue,

    /**
     * Black.
     */
    Black = 3,

    /**
     * Alpha.
     */
    Alpha = 4,

    /**
     * Index.
     */
    Index = 5,

    /**
     * Meta 0.
     */
    Meta0 = 10,

    /**
     * Meta 1.
     */
    Meta1 = 11,

    /**
     * Meta 2.
     */
    Meta2 = 12,

    /**
     * Meta 3.
     */
    Meta3 = 13,

    /**
     * Meta 4.
     */
    Meta4 = 14,

    /**
     * Meta 5.
     */
    Meta5 = 15,

    /**
     * Meta 6.
     */
    Meta6 = 16,

    /**
     * Meta 7.
     */
    Meta7 = 17,

    /**
     * Meta 8.
     */
    Meta8 = 18,

    /**
     * Meta 9.
     */
    Meta9 = 19,

    /**
     * Meta 10.
     */
    Meta10 = 20,

    /**
     * Meta 11.
     */
    Meta11 = 21,

    /**
     * Meta 12.
     */
    Meta12 = 22,

    /**
     * Meta 13.
     */
    Meta13 = 23,

    /**
     * Meta 14.
     */
    Meta14 = 24,

    /**
     * Meta 15.
     */
    Meta15 = 25,

    /**
     * Meta 16.
     */
    Meta16 = 26,

    /**
     * Meta 17.
     */
    Meta17 = 27,

    /**
     * Meta 18.
     */
    Meta18 = 28,

    /**
     * Meta 19.
     */
    Meta19 = 29,

    /**
     * Meta 20.
     */
    Meta20 = 30,

    /**
     * Meta 21.
     */
    Meta21 = 31,

    /**
     * Meta 22.
     */
    Meta22 = 32,

    /**
     * Meta 23.
     */
    Meta23 = 33,

    /**
     * Meta 24.
     */
    Meta24 = 34,

    /**
     * Meta 25.
     */
    Meta25 = 35,

    /**
     * Meta 26.
     */
    Meta26 = 36,

    /**
     * Meta 27.
     */
    Meta27 = 37,

    /**
     * Meta 28.
     */
    Meta28 = 38,

    /**
     * Meta 29.
     */
    Meta29 = 39,

    /**
     * Meta 30.
     */
    Meta30 = 40,

    /**
     * Meta 31.
     */
    Meta31 = 41,

    /**
     * Meta 32.
     */
    Meta32 = 42,

    /**
     * Meta 33.
     */
    Meta33 = 43,

    /**
     * Meta 34.
     */
    Meta34 = 44,

    /**
     * Meta 35.
     */
    Meta35 = 45,

    /**
     * Meta 36.
     */
    Meta36 = 46,

    /**
     * Meta 37.
     */
    Meta37 = 47,

    /**
     * Meta 38.
     */
    Meta38 = 48,

    /**
     * Meta 39.
     */
    Meta39 = 49,

    /**
     * Meta 40.
     */
    Meta40 = 50,

    /**
     * Meta 41.
     */
    Meta41 = 51,

    /**
     * Meta 42.
     */
    Meta42 = 52,

    /**
     * Meta 43.
     */
    Meta43 = 53,

    /**
     * Meta 44.
     */
    Meta44 = 54,

    /**
     * Meta 45.
     */
    Meta45 = 55,

    /**
     * Meta 46.
     */
    Meta46 = 56,

    /**
     * Meta 47.
     */
    Meta47 = 57,

    /**
     * Meta 48.
     */
    Meta48 = 58,

    /**
     * Meta 49.
     */
    Meta49 = 59,

    /**
     * Meta 50.
     */
    Meta50 = 60,

    /**
     * Meta 51.
     */
    Meta51 = 61,

    /**
     * Meta 52.
     */
    Meta52 = 62,

    /**
     * Composite.
     */
    Composite = 64,
}